Initial HTML plus some basic css
The HTML for the form is included, along with some css to handle the image down the left side of the form area.
This section looks good, however the right hand side where the majority of the text and the actual form are is largely unstyled.

I am committing to github and setting up a github pages page mainly to check how the current styles appear on another device with a smaller screen resolution.